requirements.txt
asn1crypto==0.24.0
beautifulsoup4==4.6.3
bottle==0.12.13
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
conda==4.5.4
cryptography==2.2.2
future==0.17.1
idna==2.6
lxml==4.2.5
numpy==1.15.4
pandas==0.23.4
pycosat==0.6.3
pycparser==2.18
pyOpenSSL==18.0.0
PySocks==1.6.8
python-dateutil==2.7.5
python-telegram-bot==11.1.0
pytz==2018.7
requests==2.18.4
ruamel-yaml==0.15.37
six==1.11.0
Unidecode==1.0.22
urllib3==1.22
